requirements.txt
attrs==17.4.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
djangorestframework==3.8.2
future==0.16.0
idna==2.6
more-itertools==4.1.0
oauthlib==2.0.7
pkg-resources==0.0.0
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
pytest==3.5.0
pytest-django==3.2.1
python-twitter==3.4.1
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
